all warnings removed from all components of the DLL. Update mylibsofa required.

+++ NativeCode/Plugin_ImpactGenerator.cpp
@@ -67,7 +67,7 @@ namespace ImpactGenerator
         RegisterParameter(definition, "Instance", "", 0.0f, MAXINSTANCES - 1, 0.0f, 1.0f, 1.0f, P_INSTANCE, "Determines the instance from which impacts are received via ImpactGenerator_AddImpact");
         RegisterParameter(definition, "Gain", "dB", -120.0f, 50.0f, 0.0f, 1.0f, 1.0f, P_GAIN, "Overall gain");
         RegisterParameter(definition, "Threshold", "dB", -120.0f, 0.0f, -60.0f, 1.0f, 1.0f, P_THR, "Threshold after which impacts stop playing");
-        RegisterParameter(definition, "MaxImpacts", "", 1.0f, MAXIMPACTS, 200.0f, 1.0f, 1.0f, P_MAXIMPACTS, "Maximum number of impact sounds played simultaneously on this instance");
+        RegisterParameter(definition, "MaxImpacts", "", 1.0f, (float)MAXIMPACTS, 200.0f, 1.0f, 1.0f, P_MAXIMPACTS, "Maximum number of impact sounds played simultaneously on this instance");
         return numparams;
     }
 
@@ -97,7 +97,7 @@ namespace ImpactGenerator
         {
             ImpactInstance* instance = GetImpactInstance((int)data->p[P_INSTANCE]);
             if (instance != NULL)
-                instance->maximpacts = value;
+                instance->maximpacts = (int)value;
         }
         return UNITY_AUDIODSP_OK;
     }
